+++ test/clj_kondo/main_test.clj
@@ -1616,6 +1616,11 @@ foo/foo ;; this does use the private var
                      :level :error,
                      :message "a number is not a function"})
                   (lint! "(1 1)"))
+  (assert-submaps
+   [{:file "<stdin>", :row 1, :col 21, :level :error
+     :message "Can't call a string as a function"}]
+   (lint! "(def foo \"foo\") foo (foo 1) (defn bar [x] x) (bar foo)"
+          {:linters {:type-mismatch {:level :error}}}))
   (is (empty? (lint! "'(1 1)")))
   (is (empty? (lint! "(foo (1 1))" "--config"
                      "{:linters {:not-a-function {:skip-args [user/foo]}